+++ packages/arcgis-rest-developer-credentials/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @esri/arcgis-rest-developer-credentials Changelog
 
+## 4.11.0
+
+### Patch Changes
+
+- Updated dependencies [05b9926]
+  - @esri/arcgis-rest-portal@4.11.0
+  - @esri/arcgis-rest-request@4.11.0
+
+## 4.10.3
+
+### Patch Changes
+
+- Updated dependencies [412188c]
+  - @esri/arcgis-rest-request@4.10.3
+  - @esri/arcgis-rest-portal@4.10.3
+
+## 4.10.2
+
+### Patch Changes
+
+- Updated dependencies [4cce28a]
+  - @esri/arcgis-rest-portal@4.10.2
+  - @esri/arcgis-rest-request@4.10.2
+
 ## 4.10.1
 
 ### Patch Changes